Rename the reducers.
