+++ src/store/index.ts
@@ -1,5 +1,5 @@
 import { configureStore } from "@reduxjs/toolkit";
-import rootReducers from "./reducers"; // 引入 reducer 的集合
+import rootReducers from "./reducer";
 
 // 实例化 store，全局唯一
 const store = configureStore({
