+++ src/store/reducer.ts
@@ -1,0 +1,8 @@
+import { combineReducers } from '@reduxjs/toolkit';
+import authSlice from './auth/authSlice';
+
+const rootReducers = combineReducers({
+  auth: authSlice
+})
+
+export default rootReducers
